+++ src/app.rs
@@ -117,6 +117,9 @@ pub async fn app(open_file: Option<PathBuf>, backend: Backend) -> IdiomResult<()
                     term.resize(gs.editor_area.width as u16);
                 }
                 Event::Mouse(event) => gs.map_mouse(event, &mut tree, &mut workspace),
+                Event::Paste(clip) => {
+                    gs.passthrough_paste(clip, &mut workspace, &mut term);
+                }
                 _ => (),
             }
         }